requirements.txt
python-dotenv==0.21.1
aiogram==2.25.1
gino==1.0.1
SQLAlchemy==1.3.24
asyncpg==0.27.0